- added Softmax
- changed function to async for llama2-wasm example

+++ candle-wasm-examples/llama2-c/src/model.rs
@@ -50,7 +50,7 @@ impl Cache {
             Ok(mask.clone())
         } else {
             let mask: Vec<_> = (0..t)
-                .flat_map(|i| (0..t).map(move |j| u8::from(j > i)))
+                .flat_map(|i| (0..t).map(move |j| u32::from(j > i)))
                 .collect();
             let mask = Tensor::from_slice(&mask, (t, t), &self.device)?;
             masks.insert(t, mask.clone());
